This should have been checked in with vtkDesktopDeliveryClient.h  "Overload of a virtual method with a different type for argument"

+++ Servers/Filters/vtkDesktopDeliveryClient.cxx
@@ -284,7 +284,7 @@ void vtkDesktopDeliveryClient::ComputeVisiblePropBounds(vtkRenderer *ren,
 }
 
 void vtkDesktopDeliveryClient
-    ::SetImageReductionFactorForUpdateRate(float DesiredUpdateRate)
+    ::SetImageReductionFactorForUpdateRate(double DesiredUpdateRate)
 {
   this->Superclass::SetImageReductionFactorForUpdateRate(DesiredUpdateRate);
   if (this->Squirt)
